feat: add mock consensus client
Introduce a mock implementation of a consensus client that fetches Ethereum block headers and publishes them to the execution layer. The mock client performs basic syncing by retrieving headers from the genesis to the latest block on startup, and subscribing to new block headers thereafter.

Later, we plan on replacing the mock implementation with a consensus client that implements the Altair Light Client protocol.