fix: retrieve touched accounts using `debug_traceTransaction` RPC (#1)
The RPC method `eth_createAccessList` was initially used to analyze
which storage slots and accounts are access during a transaction.
However, it proved unreliable for transactions that have already been
executed on-chain, since it simulates the access list based on the
current chain state, not the state at the time of execution. To address
this, we now use `debug_traceTransaction` with a `prestateTracer`.